pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crisprscope"
version = "0.1.0"
authors = [
  { name="Chetan Elenki", email="chetan.elenki@gmail.com" },
]
description = "A package for processing and analyzing single-cell CRISPR screen data from direct DNA sequencing."
readme = "README.md"
requires-python = ">=3.9"
keywords = ["crispr", "anndata", "bioinformatics", "single-cell", "genomics", "crispresso"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "anndata~=0.10.0",
    "pandas~=2.2.0",    # Pinned to a recent, stable major version
    "numpy~=1.26.0",    # Pinned to a recent, stable major version
    "h5py~=3.10.0",
    "scipy~=1.11.0",
    "pyyaml~=6.0",      # Added for config file parsing
    "dnaio~=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "jupyter",
    "build",
    "twine",
]

[project.urls]
"Homepage" = "https://github.com/elenki/crisprscope"
"Bug Tracker" = "https://github.com/elenki/crisprscope/issues"

[project.scripts]
crisprscope = "crisprscope.cli:main"